requirements.txt
alembic==1.0.3
astroid==2.0.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mfrc522==0.0.4
paho-mqtt==1.4.0
pkg-resources==0.0.0
pylint==2.1.1
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
python-engineio==3.5.0
python-socketio==3.1.2
requests==2.21.0
RPi.GPIO==0.6.5
six==1.11.0
SPI-Py==1.0
spidev==3.4
SQLAlchemy==1.2.14
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
